Refactor: follow단 api 파일 병합 (#285)
