+++ src/api/follow.js
@@ -1,40 +1,39 @@
-// const followAPI = async (accountname, token) => {
-//   try {
-//     const follow_URL = `https://api.mandarin.weniv.co.kr/profile/${accountname}/follow`
-//     const response = await fetch(follow_URL, {
-//       method: 'POST',
-//       headers: {
-//         "Authorization": `Bearer ${token}`,
-//         "Content-type": "application/json"
-//       }
-//     })
+import { authInstance } from "./instance";
 
-//     const data = await response.json();
-//     return data;
-//   } catch (error) {
-//     console.error('Account API 에러가 발생했습니다', error);
-//   }
-// };
-
-// export default followAPI;
-
-import axios from 'axios';
-
-const followAPI = async (accountname, token) => {
+export const followAPI = async (accountname, token) => {
   try {
-    const follow_URL = `https://api.mandarin.weniv.co.kr/profile/${accountname}/follow`;
-    const response = await axios.post(follow_URL, {}, {
-      headers: {
-        'Authorization': `Bearer ${token}`,
-        'Content-type': 'application/json',
-      },
-    });
-
+    const response = await authInstance.post(`/profile/${accountname}/follow`)
     return response.data;
   } catch (error) {
-    console.error('Account API 에러가 발생했습니다', error);
     throw error;
   }
 };
 
-export default followAPI;
+export const unfollowAPI = async (accountname, token) => {
+  try {
+    const response = await authInstance.delete(`/profile/${accountname}/unfollow`)  
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+export const followerAPI = async (accountname, token) => {
+  try {
+    const response = await authInstance.get(`/profile/${accountname}/follower`)
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+export const followingAPI = async (accountname, token) => {
+  try {
+    const response = await authInstance.get(`/profile/${accountname}/following`)
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+
